# openapi.yaml
openapi: 3.0.3
info:
  title: Aframp Pay API
  version: 0.1.0
  description: >
    Merchant-facing backend for Aframp Pay — Stellar-settled point-of-sale payments.
    Prose reference with frontend guidance lives in API.md; this spec is the
    machine-readable contract for generating typed clients.


    Money is always integer stroops (1 unit = 10,000,000 stroops), never a float.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /payment-requests/{id}:
    get:
      tags: [Payment requests]
      summary: Read a payment request (public)
      description: >
        Deliberately unauthenticated so a customer's device can read a request
        before paying. Prefer GET /payment-requests/{id}/status for polling
        after the customer has submitted payment.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Payment request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        HS256, 24h expiry. No refresh endpoint — a 401 means re-login. For API
        clients and scripts; browsers should use cookieAuth instead.
    cookieAuth:
      type: apiKey
      in: cookie
      name: aframp_session
      description: >
        HttpOnly session cookie set by /signup and /login, cleared by /logout.
        The preferred method for browsers — JS cannot read it, so an XSS on the
        page cannot steal the session. Never copy the token into localStorage.
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Same JWT as bearerAuth, but the token must have been issued to a user
        with `is_admin = true`. The flag is baked into the JWT at login — flip
        the DB flag and re-login to pick up the change.
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: aframp_session
      description: >
        Same session cookie, but the JWT must carry `is_admin = true`.
        Re-login after flipping `is_admin` in Postgres.
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    ApiError:
      type: object
      required: [error, code]
      properties:
        error: { type: string, example: insufficient available balance }
        code:
          type: string
          example: INSUFFICIENT_BALANCE
          description: "Machine-readable error code. Full catalog: INVALID_PARAMETERS, INVALID_AMOUNT, INSUFFICIENT_BALANCE, UNSUPPORTED_ASSET, PAYOUT_FAILED, EMAIL_TAKEN, PHONE_TAKEN, INVALID_CREDENTIALS, USER_NOT_FOUND, MERCHANT_NOT_FOUND, WALLET_NOT_FOUND, PAYMENT_REQUEST_NOT_FOUND, FORBIDDEN, OTP_INVALID, OTP_EXPIRED, OTP_LOCKED, OTP_CHALLENGE_NOT_FOUND, TOO_MANY_REQUESTS, INTERNAL_ERROR"
...
    PaymentRequest:
      type: object
      required: [id, merchant_id, address, network, amount_stroops, asset, memo, status, expires_at, created_at]
      properties:
        id: { type: string, format: uuid }
        merchant_id: { type: string, format: uuid }
        address: { type: string, description: Stellar destination address to pay. }
        network: { type: string, example: stellar }
        amount_stroops: { type: integer, format: int64 }
        asset: { type: string, example: XLM }
        memo:
          type: string
          description: >
            Correlation key. A payment without this memo still credits the balance
            but leaves the request pending forever. Included in sep7_uri automatically.
        status:
          type: string
          enum: [pending, paid, expired]
          description: >
            'expired' is computed at read time from expires_at. A request that
            expires and is then paid still flips to 'paid'.
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        sep7_uri:
          type: string
          nullable: true
          description: >
            SEP-0007 payment URI to render as a QR code. Null for non-XLM assets —
            no cNGN issuer address is configured, and a guessed one would misdirect funds.
          example: web+stellar:pay?destination=GDDT...&amount=2.5000000&memo=1f97c93409172a7d&memo_type=MEMO_TEXT
...